upload/requirements.txt
pyrebase4
pycryptodome